fix : array_size type change

+++ 07/ex02/Array.hpp
@@ -6,7 +6,7 @@
 /*   By: dohelee <dohelee@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/11 04:02:28 by dohelee           #+#    #+#             */
-/*   Updated: 2021/10/11 04:02:28 by dohelee          ###   ########.fr       */
+/*   Updated: 2021/10/12 23:11:37 by dohelee          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,8 +19,8 @@ template<class T>
 class Array
 {
 	private:
-		T*	arr;
-		int	array_size;
+		T*				arr;
+		unsigned int	array_size;
 	public:
 		Array()
 		{
@@ -50,7 +50,7 @@ class Array
 			{
 				this->arr = new T[op.size()]();
 				this->array_size = op.size();
-				for (int i = 0; i < op.size(); i++)
+				for (unsigned int i = 0; i < op.size(); i++)
 				{
 					this->arr[i] = op.arr[i];
 				}
@@ -60,13 +60,13 @@ class Array
 
 		T& operator[](unsigned int i)
 		{
-			if (i > static_cast<unsigned int>(this->size()) - 1)
+			if (i > this->size() - 1)
 				throw ArrayException();
 			else
 				return (this->arr[i]);
 		};
 
-		int size(void) const
+		unsigned int size(void) const
 		{
 			return (this->array_size);
 		};
